Add files via upload
The three milestones that have been provided is the various progress throughout the entire course from the base code to the enhancements per milestone. Each zip file also has an associated reflection to go with it that was provided as an overview of what has taken place during that week.